Files have been modified

+++ ApexRoot/1ApexCommon/src/main/java/com/nus/fileupload/entities/HprojectMonthlyResourceUtilization.java
@@ -32,10 +32,10 @@ public class HprojectMonthlyResourceUtilization implements Serializable {
 
     //--- ENTITY DATA FIELDS 
     @Column(name="billed_days")
-    private Integer    billedDays ;
+    private Double    billedDays ;
 
     @Column(name="available_days")
-    private Integer    availableDays ;
+    private Double    availableDays ;
 
     @Column(name="project_code_id", nullable=false)
     private Integer    projectCodeId ;
@@ -100,17 +100,17 @@ public class HprojectMonthlyResourceUtilization implements Serializable {
         return this.id;
     }
 
-    public void setBilledDays( Integer billedDays ) {
+    public void setBilledDays( Double billedDays ) {
         this.billedDays = billedDays ;
     }
-    public Integer getBilledDays() {
+    public Double getBilledDays() {
         return this.billedDays;
     }
 
-    public void setAvailableDays( Integer availableDays ) {
+    public void setAvailableDays( Double availableDays ) {
         this.availableDays = availableDays ;
     }
-    public Integer getAvailableDays() {
+    public Double getAvailableDays() {
         return this.availableDays;
     }
